--- handle/type.go
@@ -1,1 +1,0 @@
-package handle
